Add a checkbox for toggling opacity in 3d redering
By default, regions with higher density region have higher opacity. In the
current implementation, there is no option to change this behavior. At
least, I cannot find it. So I added a checkbox for inverting the opacity
mapping. This is useful when the high-density region is outside the main
volume.